Solving a business problem with RFM analysis
FLO store is putting up a new brand of women's shoes for sale. The new brand of the store is above the general customer preferences. For this reason, it is requested to contact customers in the profile who may be interested in the promotion of the brand and the sale of products. Shoppers from loyal customers (champions, loyal_customers) and women's categories are customers who need to be contacted in particular. The ID numbers of these customers will be forwarded to the sales and marketing department with a CSV file.